Add zsh-help function for easily searching the zsh documentation (#1360)
* Add zsh-help function for easily searching the zsh documentation

Looks up things in the zsh documentation.
Usage: zsh-help [--all] search term(s)

Option --all will seach for the term anywhere, not just at the start of a
line. When not using --all it will search nicely for terms at the beginning
of the line, which in the zsh man pages is where terms that are explained
are located, allowing you to search the zsh man pages easily.

* Improve zsh-help to search section headings before other text

Provides a much easier way to search and access ZSH's manual. First checks for
terms at the start of the manual, then checks if it's at start of a line allowing
whitespace.

Clean up some of the code a bit and format it to have a proper header for the
zprezto project with author/email and description of the function.